car_info/requirements.txt
Django>=3.0,<4.0
djangorestframework
drf-spectacular
beautifulsoup4
lxml
requests
